requirements.txt
django==1.10.5
jinja2==2.9.5
djangorestframework==3.5.3
